RENAME domain으로 패키지명 변경
