+++ src/main/java/com/jeontongju/authentication/repository/SnsAccountRepository.java
@@ -1,6 +1,6 @@
 package com.jeontongju.authentication.repository;
 
-import com.jeontongju.authentication.entity.SnsAccount;
+import com.jeontongju.authentication.domain.SnsAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
