+++ src/main/java/com/jeontongju/authentication/service/MemberService.java
@@ -9,9 +9,8 @@ import com.jeontongju.authentication.dto.response.oauth.google.GoogleOAuthInfo;
 import com.jeontongju.authentication.dto.response.oauth.kakao.KakaoOAuthInfo;
 import com.jeontongju.authentication.dto.temp.ConsumerInfoForCreateBySnsRequestDto;
 import com.jeontongju.authentication.dto.temp.MemberEmailForKeyDto;
-import com.jeontongju.authentication.dto.temp.SellerInfoForCreateRequestDto;
-import com.jeontongju.authentication.entity.Member;
-import com.jeontongju.authentication.entity.SnsAccount;
+import com.jeontongju.authentication.domain.Member;
+import com.jeontongju.authentication.domain.SnsAccount;
 import com.jeontongju.authentication.enums.MemberRoleEnum;
 import com.jeontongju.authentication.enums.SnsTypeEnum;
 import com.jeontongju.authentication.exception.*;
@@ -34,7 +33,6 @@ import io.jsonwebtoken.security.Keys;
 import io.jsonwebtoken.security.SignatureException;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
-import java.util.Optional;
 import javax.crypto.SecretKey;
 import javax.mail.MessagingException;
 import javax.persistence.EntityNotFoundException;
